Feature|MapMaker: Implemented token scaling and panning

+++ map-generator/RenderPipeline/AvaloniaRenderPipeline.cs
@@ -81,15 +81,15 @@ public class AvaloniaRenderPipeline : AbstractRenderPipeline
     /**
      * Converts a location on the WriteableBitmap into a position on the map.
      */
-    public (float x, float y) ScreenToWorldspace(float x, float y, float zoom, (float x, float y) screenPosition)
+    public (double x, double y) ScreenToWorldspace(double x, double y, float zoom, (double x, double y) screenPosition)
     {
         (float tilesX, float tilesY) = CalculateConstraints(zoom);
 
-        float tlX = x - tilesX / 2;
-        float tlY = y - tilesY / 2;
+        double tlX = x - tilesX / 2;
+        double tlY = y - tilesY / 2;
 
-        float widthRatio = screenPosition.x / _bitmapWidth;
-        float heightRatio = screenPosition.y / _bitmapHeight;
+        double widthRatio = screenPosition.x / _bitmapWidth;
+        double heightRatio = screenPosition.y / _bitmapHeight;
 
         return (
             tlX + tilesX * widthRatio,
@@ -100,12 +100,11 @@ public class AvaloniaRenderPipeline : AbstractRenderPipeline
     /**
      * Converts a position on the map into a location on the WriteableBitmap
      */
-    public (float x, float y) WorldToScreenspace(float x, float y, float zoom, (float x, float y) worldPosition)
+    public (double x, double y) WorldToScreenspace(double x, double y, float zoom, (double x, double y) worldPosition)
     {
         (float tilesX, float tilesY) = CalculateConstraints(zoom);
-
-        float tlX = x - tilesX / 2;
-        float tlY = y - tilesY / 2;
+        double tlX = x - tilesX / 2;
+        double tlY = y - tilesY / 2;
 
         return (
             (worldPosition.x - tlX) * _bitmapWidth / tilesX,
@@ -142,7 +141,7 @@ public class AvaloniaRenderPipeline : AbstractRenderPipeline
         float tilesX;
         float tilesY;
 
-        float aspectRatio = _bitmapWidth / (float)_bitmapHeight;
+        float aspectRatio = (float)_bitmapWidth / (float)_bitmapHeight;
 
         if (aspectRatio > mapAspectRatio)
         {
